% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_analysis.R
\name{analyze_runs}
\alias{analyze_runs}
\title{One-call analysis (batch)}
\usage{
analyze_runs(
  results_dir = "scripts/results",
  analysis_dir = "scripts/results/analysis",
  cmap_meta_path,
  cmap_valid_path,
  cmap_signatures_path,
  q_thresh = 0.05,
  reversal_only = TRUE,
  verbose = TRUE
)
}
\description{
One-call analysis (batch)
}
