% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_analysis.R
\name{annotate_filter_runs}
\alias{annotate_filter_runs}
\title{Annotate/filter a list of per-run drug tables using CMAP metadata}
\usage{
annotate_filter_runs(
  drugs_list,
  cmap_meta_path,
  cmap_valid_path,
  q_thresh = 0.05,
  reversal_only = TRUE
)
}
\description{
Annotate/filter a list of per-run drug tables using CMAP metadata
}
