% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_config.R
\name{io_list_disease_files}
\alias{io_list_disease_files}
\title{List disease files given a directory + pattern OR a single CSV path}
\usage{
io_list_disease_files(disease_path, disease_pattern = NULL)
}
\arguments{
\item{disease_path}{A directory or a single CSV path}

\item{disease_pattern}{Pattern to search when \code{disease_path} is a directory}
}
\description{
List disease files given a directory + pattern OR a single CSV path
}
\keyword{internal}
