% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_analysis.R
\name{load_run_results}
\alias{load_run_results}
\title{Load a set of pipeline result files (RData) produced by run_dr()}
\usage{
load_run_results(result_files)
}
\arguments{
\item{result_files}{Vector of file paths to RData files containing results}
}
\description{
Load a set of pipeline result files (RData) produced by run_dr()
}
