% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{random_score}
\alias{random_score}
\title{Generate a null distribution of reversal scores using random genes}
\usage{
random_score(cmap_signatures, n_up, n_down, N_PERMUTATIONS = 1e+05, seed = 123)
}
\arguments{
\item{cmap_signatures}{matrix/data.frame with first column = Entrez IDs,
subsequent columns = ranked values per experiment}

\item{n_up}{integer; number of up-regulated disease genes}

\item{n_down}{integer; number of down-regulated disease genes}

\item{N_PERMUTATIONS}{integer; number of random iterations (default 1e5)}

\item{seed}{integer for RNG seed (default 123)}
}
\value{
numeric vector of random (null) connectivity scores
}
\description{
Generate a null distribution of reversal scores using random genes
}
