% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_processing.R
\name{run_dr}
\alias{run_dr}
\title{Run the drug repurposing pipeline in one call}
\usage{
run_dr(
  signatures_rdata,
  disease_path,
  disease_pattern = NULL,
  cmap_meta_path = NULL,
  cmap_valid_path = NULL,
  out_dir = "scripts/results",
  gene_key = "SYMBOL",
  logfc_cols_pref = "log2FC",
  logfc_cutoff = 1,
  q_thresh = 0.05,
  reversal_only = TRUE,
  seed = 123,
  verbose = TRUE,
  make_plots = TRUE
)
}
\arguments{
\item{signatures_rdata}{Path to RData containing \code{cmap_signatures}}

\item{disease_path}{CSV path OR directory containing the disease file}

\item{disease_pattern}{Pattern to locate CSV inside a directory}

\item{cmap_meta_path}{Path to CMAP experiment metadata CSV (optional)}

\item{cmap_valid_path}{Path to CMAP valid instances CSV (optional)}

\item{out_dir}{Output directory}
}
\value{
Invisibly, the DRP object
}
\description{
Run the drug repurposing pipeline in one call
}
