% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_analysis.R
\name{report_runs}
\alias{report_runs}
\title{Make standard per-run plots & CSVs}
\usage{
report_runs(
  drugs_list,
  signatures_list,
  cmap_signatures_path,
  out_dir = "scripts/results/analysis"
)
}
\description{
Make standard per-run plots & CSVs
}
